data_collection/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
blessed==1.20.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
editor==1.6.6
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inquirer==3.4.0
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.20
langchain-core==0.3.41
langchain-text-splitters==0.3.6
langsmith==0.3.11
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
openai==1.65.4
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
praw==7.8.1
prawcore==2.4.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-googlesearch==1.1.1
pytz==2025.1
pyuseragents==1.0.5
PyYAML==6.0.2
readchar==4.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
runs==1.2.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
xmod==1.8.1
youtube-transcript-api==1.0.0
zstandard==0.23.0
